+++ ODMS/src/main/java/org/example/Catalog/CatalogRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CatalogRepository extends JpaRepository<Catalog, Integer> {
+public interface CatalogRepository extends JpaRepository<Catalog, Long> {
    List<Catalog> getCatalogByCategory(Category category);
    List<Catalog> findByCatalogNameContainingIgnoreCase(String catalogName);
 }
