+++ ODMS/src/main/java/org/example/OrderItems/OrderItemsRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OrderItemsRepository extends JpaRepository<OrderItems, Integer> {
+public interface OrderItemsRepository extends JpaRepository<OrderItems, Long> {
     List<OrderItems> findByOrders(Orders orders);
     void deleteByOrders(Orders orders);
 }
