+++ ODMS/src/main/java/org/example/Users/Users.java
@@ -16,7 +16,7 @@ import java.util.Set;
 public class Users {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer userId;
+    private Long userId;
 
     private String username;
     private String password;
@@ -34,7 +34,7 @@ public class Users {
 
 
 
-    public Users(Integer userId, String username, String password, Boolean enabled, String role) {
+    public Users(Long userId, String username, String password, Boolean enabled, String role) {
         this.userId = userId;
         this.username = username;
         this.password = password;
@@ -46,11 +46,11 @@ public class Users {
 
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
